feat: :sparkles: 전체 컴포넌트 정리
